# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SupervIA - AI Service API",
    "version": "1.0.0",
    "description": "API du service d'intelligence artificielle de SupervIA. Il expose deux agents principaux : le **Dashboard Agent** pour la génération de tableaux de bord, et l'**Ops Assistant** pour le diagnostic et l'assistance opérationnelle via un chat interactif.",
    "contact": {
      "name": "SupervIA Team",
      "url": "https://github.com/zkerkeb-class/service-ia-Tadayoshi123"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3004",
      "description": "Serveur de développement local"
    },
    {
      "url": "http://localhost:8080/api/v1/ai",
      "description": "Via la passerelle API Nginx"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token d'accès JWT (RS256) pour l'authentification des utilisateurs finaux."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "nullable": true
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "DashboardConfig": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre du dashboard."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Description du dashboard."
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardBlock"
            }
          }
        },
        "required": [
          "title",
          "blocks"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/dashboard-agent/generate": {
      "post": {
        "summary": "Génère un dashboard complet",
        "tags": [
          "Dashboard Agent"
        ],
        "description": "Crée une configuration de dashboard JSON complète basée sur les exigences de l'utilisateur, les métriques disponibles et des templates optionnels.",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDashboardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dashboard généré avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "dashboard": {
                      "$ref": "#/components/schemas/DashboardConfig"
                    },
                    "recommendations": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Suggestions pour améliorer le dashboard."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreurs de validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard-agent/suggest-layout": {
      "post": {
        "summary": "Suggère une disposition de dashboard",
        "tags": [
          "Dashboard Agent"
        ],
        "description": "Propose une disposition de grille optimale pour un ensemble de métriques donné.",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "metrics": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Liste des noms de métriques à inclure.",
                    "example": [
                      "cpu_usage",
                      "memory_usage",
                      "api_latency"
                    ]
                  },
                  "screenSize": {
                    "type": "string",
                    "enum": [
                      "mobile",
                      "tablet",
                      "desktop",
                      "large"
                    ],
                    "description": "La taille d'écran cible pour l'optimisation.",
                    "example": "desktop"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "performance",
                      "aesthetics",
                      "density"
                    ],
                    "description": "Le critère principal pour l'optimisation de la disposition.",
                    "example": "performance"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suggestion de disposition générée."
          },
          "400": {
            "description": "Erreurs de validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard-agent/optimize": {
      "post": {
        "summary": "Optimise un dashboard existant",
        "tags": [
          "Dashboard Agent"
        ],
        "description": "Analyse une configuration de dashboard existante et propose des optimisations pour améliorer les performances et l'expérience utilisateur.",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentDashboard": {
                    "type": "object",
                    "description": "La configuration JSON du dashboard à optimiser.",
                    "example": {
                      "dashboard": {
                        "title": "Mon Dashboard",
                        "blocks": [
                          {
                            "id": "block1",
                            "type": "metric",
                            "title": "CPU Usage"
                          }
                        ]
                      }
                    }
                  },
                  "optimizationGoals": {
                    "type": "object",
                    "description": "Objectifs d'optimisation spécifiques (optionnel).",
                    "example": {
                      "performance": true,
                      "usability": true,
                      "readability": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Optimisations proposées."
          },
          "400": {
            "description": "Erreurs de validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard-agent/validate": {
      "post": {
        "summary": "Valide une configuration de dashboard",
        "tags": [
          "Dashboard Agent"
        ],
        "description": "Valide la structure et la cohérence d'une configuration de dashboard avec des recommandations d'amélioration.",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dashboardConfig": {
                    "type": "object",
                    "description": "Configuration du dashboard à valider",
                    "example": {
                      "dashboard": {
                        "title": "Dashboard Test",
                        "blocks": [
                          {
                            "id": "metric1",
                            "type": "metric",
                            "title": "CPU Usage",
                            "layout": {
                              "x": 0,
                              "y": 0,
                              "w": 6,
                              "h": 4
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation terminée avec succès."
          },
          "400": {
            "description": "Erreurs de validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard-agent/suggest-templates": {
      "post": {
        "summary": "Suggère des templates de dashboard",
        "tags": [
          "Dashboard Agent"
        ],
        "description": "Suggère des templates appropriés selon le cas d'usage et les préférences utilisateur.",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "useCase": {
                    "type": "string",
                    "description": "Cas d'usage du dashboard",
                    "example": "Surveillance des performances d'une application web",
                    "minLength": 5,
                    "maxLength": 200
                  },
                  "complexity": {
                    "type": "string",
                    "enum": [
                      "simple",
                      "medium",
                      "complex"
                    ],
                    "description": "Complexité souhaitée",
                    "example": "medium"
                  },
                  "preferences": {
                    "type": "object",
                    "description": "Préférences utilisateur (optionnel)",
                    "example": {
                      "focus": "performance",
                      "style": "modern",
                      "metrics": [
                        "cpu",
                        "memory",
                        "latency"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suggestions générées avec succès."
          },
          "400": {
            "description": "Erreurs de validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard-agent/recommend-blocks": {
      "post": {
        "summary": "Recommande des blocs pour un dashboard",
        "tags": [
          "Dashboard Agent"
        ],
        "description": "Suggère des blocs supplémentaires pour améliorer un dashboard existant selon son objectif.",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dashboardPurpose": {
                    "type": "string",
                    "description": "Objectif du dashboard",
                    "example": "Surveillance des performances d'une API",
                    "minLength": 5,
                    "maxLength": 200
                  },
                  "existingBlocks": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "Blocs existants dans le dashboard",
                    "example": [
                      {
                        "id": "cpu-block",
                        "type": "metric",
                        "title": "CPU Usage"
                      }
                    ]
                  },
                  "focus": {
                    "type": "string",
                    "description": "Focus du dashboard (optionnel)",
                    "example": "performance"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recommandations générées avec succès."
          },
          "400": {
            "description": "Erreurs de validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard-agent/templates": {
      "get": {
        "summary": "Récupère les templates disponibles",
        "tags": [
          "Dashboard Agent"
        ],
        "description": "Retourne la liste des templates de dashboard disponibles avec filtrage optionnel.",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filtrer par catégorie de template",
            "example": "infrastructure"
          },
          {
            "in": "query",
            "name": "complexity",
            "schema": {
              "type": "string",
              "enum": [
                "simple",
                "medium",
                "complex"
              ]
            },
            "description": "Filtrer par niveau de complexité",
            "example": "medium"
          }
        ],
        "responses": {
          "200": {
            "description": "Templates récupérés avec succès."
          },
          "500": {
            "description": "Erreur interne du serveur.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}